feat: add functions to expose slots
Problem: we need to quickly get slots to schedule
Solution: provide functions to do so. If a jobspec
resources does not define slots, all are considered to
be slots. If the Jobspec does not define slots (or does)
but some subset are marked for scheduling, filter to
that set. Two different functions are provided to
provide either.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>